warning when light unavilable (#2771)
* error-card when light unavilable

* single warning element for all

* address review comments

* address review comments
